requirements.txt
asgiref==3.5.2
Django==4.0.4
django-dotenv==1.4.2
djangorestframework==3.13.1
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
algoliasearch-django>=2.0,<3.0